pyproject.toml
[project]
name = "netbox-super-cli"
version = "1.0.4"
description = "Dynamic NetBox CLI generated from the live OpenAPI schema."
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [{ name = "Thomas", email = "mick27@gmail.com" }]
keywords = ["netbox", "cli", "openapi", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "httpx>=0.27",
    "pydantic>=2.7",
    "ruamel.yaml>=0.18",
    "structlog>=24.1",
    "platformdirs>=4.2",
]

[project.scripts]
netbox-super-cli = "nsc.cli.app:main"
nsc = "nsc.cli.app:main"

[project.urls]
Homepage = "https://github.com/thomaschristory/netbox-super-cli"
Issues = "https://github.com/thomaschristory/netbox-super-cli/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nsc"]

[tool.hatch.build.targets.wheel.force-include]
# `nsc/schemas/bundled/` is inside the `nsc` package and already ships
# via `packages = ["nsc"]` above. Listing it here too produced wheels
# with duplicate local-file headers, which PyPI's archive validator
# rejects with a 400 (see https://docs.pypi.org/archives). Only
# `skills/netbox-super-cli/`, which lives outside the package tree,
# genuinely needs force-include.
"skills/netbox-super-cli" = "skills/netbox-super-cli"

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "mypy>=1.20",
    "ruff>=0.15",
    "pre-commit>=3.7",
    "types-PyYAML>=6.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "pyyaml>=6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E", "F", "W",       # pycodestyle / pyflakes
    "I",                 # isort
    "B",                 # bugbear
    "UP",                # pyupgrade
    "RUF",               # ruff-specific
    "SIM",               # simplify
    "PL",                # pylint
    "TID",               # tidy imports
    "C4",                # comprehensions
]
ignore = [
    "PLR0913",           # too many arguments — Typer commands need them
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "PLR2004",           # magic values in tests are fine (status codes, expected values)
    "PLR0915",           # e2e lifecycle tests legitimately need many statements
    "S101",              # asserts are how pytest works
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
pretty = true

[tool.pytest.ini_options]
addopts = "-ra --strict-markers --strict-config --ignore=tests/e2e"
testpaths = ["tests"]
